Fix Partial Caching fetched data calculation
Fetched data calculation was failing for files which size is
a multiple of the block size (4096).

+++ coda-src/venus/vproc.h
@@ -237,7 +237,7 @@ class vproc : public olink {
     void symlink(struct venus_cnode *, char *, struct coda_vattr *, char *);
     void readlink(struct venus_cnode *, struct coda_string *);
     void fsync(struct venus_cnode *);
-    void read(struct venus_cnode * node, uint pos, int count);
+    void read(struct venus_cnode * node, uint64_t pos, int64_t count);
 
     /* Pathname translation. */
     int namev(char *, int, struct venus_cnode *);
